변수 네이밍 변경, 함수 분리
