+++ src/pages/quiz/[id].tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import classnames from 'classnames';
-import readFile from '@lib/quizFileRead/file-read';
+import readAllQuizFile, { readQuizFileById } from '@lib/quiz/read';
 import Header from '@component/header';
 import Canvas from '@component/Canvas';
 import Editor from '@component/Editor';
@@ -82,7 +82,7 @@ export default function MainQuiz({ id, name, userHTML, userCSS, answerHTML, answ
 }
 
 export async function getStaticPaths() {
-  const paths = readFile(true)
+  const paths = readAllQuizFile()
   return {
     paths,
     fallback: false
@@ -91,7 +91,7 @@ export async function getStaticPaths() {
 
 export async function getStaticProps({ params }) {
   const index = Number(params.id).toString()
-  const quizFileData = readFile(false, index)
+  const quizFileData = readQuizFileById(index)
   return {
     props: {
       ...quizFileData
